Add files via upload
The "absenteeism-in-the-workplace" folder is set up to recreate the original database since it couldn’t be downloaded directly. Inside, there are two subfolders: imported_data_files and query. The imported_data_files folder contains three CSV files—absenteeism_at_work.csv, compensation.csv, and reasons.csv—which hold all the original data. The query folder contains an SQL file that connects, cleans, and formats these CSV files, similar to how they were structured in the database. There's also a Power BI dashboard file that brings everything together for analysis and visualization. This setup keeps the data organized and ensures it can be used in Power BI just like it was when stored in the database.